WORLD GEOMETRY IS ON SCREEN -- Paradise City draws through the real chain
Terrain, road network and buildings, drawn end-to-end: PVS zone query ->
TRK_UNIT stream -> FixUp -> instance list -> dispatch list 11 -> D3D9.
128 renderable objects -> 187 mesh draws live in the world-opaque list.
Frames: scratch/FIRST_WORLD_PIXELS/.

BLOCKER SOLVED -- RwRenderableResourceType::FixUp @0x828A9498 (the graph was
never rebased, so no mesh was walkable). The keystone was a3: it is NOT the
pool Entry's rw::Resource. Pool::FixUpEntry @0x828EB860 builds a 5-word block
on its own stack {small[0], 0, small[1], 0, 0}, so a3[2] (the GPU buffer-base
delta) is bundle MEMORY TYPE 1, not 2 -- confirmed against the data (every
type-12 entry is [hdr, body, 0]). The body rebases mppMeshes (+0x14), the
ObjectScopeTextureInfo block (+0x18) with its three tables and every
texture-name pointer, then each mesh pointer + FixUpRenderableMesh, then sets
flag bit 8. The X360 asm, renderable_transcode.py's emitted layout and the
committed consumer now all agree; RwRenderableFixUpData retyped to
{uintptr_t muPointerBase; u32 muBufferOffset;}.

PRODUCER SEAM: WorldModule::GenerateDispatchLists existed but was never
called -- BrnGameModule::DoDispatch @0x823DC458 was a stub. Landed a
documented [FLAG PC bring-up] producer (DoDispatch ->
GenerateDispatchListsBringUp -> WorldEntityModule::
GenerateDispatchListsFromStreamer -> the REAL RenderInstance) driven from
BrnMain between UpdateThread() and OnEndOfUpdateFrame().

SEVEN REAL BUGS found on the way down:
1. Pool::ResolveImportForEntry wrote 8 bytes into 4-byte serialised slots,
   wiping Model::mpu8StateRenderableIndices. Now width-aware (target above
   4GB -> full store; below -> low half, correct for widened slots because
   import slots are serialised-null).
2. Model::GetNumLods/GetNumRenderables were `return 0` stubs -> "Model in
   unit has no lods!" on every instance. Bodied in CgsModel.cpp.
3. ALL SIX command-id checks in CgsDispatcherCommands.cpp read word0 & 0x7F
   instead of (word0 >> 24) & 0x7F.
4. ShaderConstantTable's ctor was empty and all four SetShaderConstantData
   overloads were inert -- no constant could ever be set. Reconstructed the
   ctor @0x827EDDC8, AddShaderConstant @0x823F4260, and the Vector4
   @0x822B32E8 / Matrix44Affine @0x822B33B8 setters; BrnRendererModule::
   Construct now registers the 27 game constants in the X360's exact order.
5. MaterialAssembly was host-width -- GetLength() read +0x0C instead of
   +0x08. Pinned to the console 28-byte form with Ptr32<T>.
6. The no-SHADERS gate called AllocateMemoryFast inside an open packet.
7. WorldDraw_IndexedUP read vertex data 3 bytes off (the low-2 memory-class
   bits of muBaseAddress).
Plus: the world mesh-list ids were the CAR ones (19/20) and the AddToBin call
site had technique/transparent-list swapped -- corrected to 11/15/21 and
re-verified live.

Build exit 0; boot 123 frames, title cue 19.6s, 0 asserts; lint PASS on all
17 changed files. New capture driver scratch/world_shot.ps1 (PrintWindow
only). The world is UNTEXTURED: SHADERS.BNDL is converted but unstaged
(GamePrepare returns true immediately on PC) and WORLDTEX still hits the
CreateTexture(fmt=0) seam; the loading screen's opaque background covers the
world in the normal flow (hence the BRN_WORLD_ONLY=1 diagnostic).

+++ src/GameShared/GameClasses/RenderWare/PS3/CgsRwRenderableResourceTypePS3.cpp
@@ -27,8 +27,11 @@ namespace CgsResource
     void RwRenderableResourceType::FixUpRenderableMesh(RwRenderableMesh* lpMesh,
                                                        const RwRenderableFixUpData* lpFixUp) const
     {
-        const u32 luPointerBase  = lpFixUp->muPointerBase;   // a3[0]
-        const u32 luBufferOffset = lpFixUp->muBufferOffset;  // a3[2]
+        // SEAM (platform-4 widened data): the serialised pointer slots are host-width, so the
+        // relocation base is too (X360: a 32-bit `lwz 0(a3)`). The buffer offset stays 32-bit --
+        // muBaseAddress is a console u32 slot resolved through the low-4 GB reservation.
+        const uintptr_t luPointerBase  = lpFixUp->muPointerBase;   // a3[0]
+        const u32       luBufferOffset = lpFixUp->muBufferOffset;  // a3[2]
 
         // 1) Relocate each serialised vertex-buffer-header pointer slot to its loaded base.
         //    SEAM: the widened buffer table holds the IB header pointer at mapBuffers[0] and
